+++ src/Language/AST/ValueNode.php
@@ -3,15 +3,15 @@
 namespace GraphQL\Language\AST;
 
 /**
-export type ValueNode = VariableNode
-| NullValueNode
-| IntValueNode
-| FloatValueNode
-| StringValueNode
-| BooleanValueNode
-| EnumValueNode
-| ListValueNode
-| ObjectValueNode
+ * export type ValueNode = VariableNode
+ * | NullValueNode
+ * | IntValueNode
+ * | FloatValueNode
+ * | StringValueNode
+ * | BooleanValueNode
+ * | EnumValueNode
+ * | ListValueNode
+ * | ObjectValueNode.
  */
 interface ValueNode
 {