+++ src/Server/Helper.php
@@ -46,7 +46,7 @@ class Helper
      *
      * @api
      */
-    public function parseHttpRequest(?callable $readRawBodyFn = null)
+    public function parseHttpRequest(callable $readRawBodyFn = null)
     {
         $method = $_SERVER['REQUEST_METHOD'] ?? null;
         $bodyParams = [];
@@ -255,7 +255,7 @@ class Helper
                 );
             }
 
-            $doc = $op->queryId !== null && $op->query === null
+            $doc = $op->queryId !== null
                 ? $this->loadPersistedQuery($config, $op)
                 : $op->query;