+++ src/Type/Definition/InputType.php
@@ -3,17 +3,17 @@
 namespace GraphQL\Type\Definition;
 
 /**
-export type InputType =
-  | ScalarType
-  | EnumType
-  | InputObjectType
-  | ListOfType<InputType>
-  | NonNull<
-      | ScalarType
-      | EnumType
-      | InputObjectType
-      | ListOfType<InputType>,
-    >;
+ * export type InputType =
+ * | ScalarType
+ * | EnumType
+ * | InputObjectType
+ * | ListOfType<InputType>
+ * | NonNull<
+ * | ScalarType
+ * | EnumType
+ * | InputObjectType
+ * | ListOfType<InputType>,
+ * >;.
  */
 interface InputType
 {