+++ src/Type/Definition/IntType.php
@@ -69,7 +69,7 @@ values. Int can represent values between -(2^31) and 2^31 - 1. ';
         return (int) $value;
     }
 
-    public function parseLiteral(Node $valueNode, ?array $variables = null): int
+    public function parseLiteral(Node $valueNode, array $variables = null): int
     {
         if ($valueNode instanceof IntValueNode) {
             $val = (int) $valueNode->value;