+++ tests/Server/QueryExecutionTest.php
@@ -47,7 +47,7 @@ final class QueryExecutionTest extends ServerTestCase
      *
      * @throws \Exception
      */
-    private function assertQueryResultEquals(array $expected, string $query, ?array $variables = null, ?string $queryId = null): ExecutionResult
+    private function assertQueryResultEquals(array $expected, string $query, array $variables = null, string $queryId = null): ExecutionResult
     {
         $result = $this->executeQuery($query, $variables, false, $queryId);
         self::assertArraySubset($expected, $result->toArray(DebugFlag::INCLUDE_DEBUG_MESSAGE));
@@ -60,7 +60,7 @@ final class QueryExecutionTest extends ServerTestCase
      *
      * @throws \Exception
      */
-    private function executeQuery(string $query, ?array $variables = null, bool $readonly = false, ?string $queryId = null): ExecutionResult
+    private function executeQuery(string $query, array $variables = null, bool $readonly = false, string $queryId = null): ExecutionResult
     {
         $op = OperationParams::create(
             [
@@ -291,7 +291,7 @@ final class QueryExecutionTest extends ServerTestCase
      *
      * @throws \Exception
      */
-    private function executePersistedQuery(string $queryId, ?array $variables = null): ExecutionResult
+    private function executePersistedQuery(string $queryId, array $variables = null): ExecutionResult
     {
         $op = OperationParams::create([
             'queryId' => $queryId,
@@ -455,15 +455,22 @@ final class QueryExecutionTest extends ServerTestCase
         self::assertSame($expected, $result->toArray());
     }
 
-    public function testExecutesQueryWhenQueryAndQueryIdArePassed(): void
+    public function testLoadsPersistedQueryWhenQueryAndQueryIdArePassed(): void
     {
         $query = /** @lang GraphQL */ '{ f1 }';
 
         $expected = [
-            'data' => ['f1' => 'f1'],
+            'errors' => [
+                [
+                    'message' => 'Cannot query field "invalid" on type "Query".',
+                    'locations' => [['line' => 1, 'column' => 3]],
+                ],
+            ],
         ];
-        $this->config->setPersistedQueryLoader(static function (): array {
-            throw new \Exception('Should not be called since a query is also passed');
+        $this->config->setPersistedQueryLoader(static function (string $queryId, OperationParams $params) use ($query): string {
+            self::assertSame($query, $params->query);
+
+            return /** @lang GraphQL */ '{ invalid }';
         });
 
         $this->assertQueryResultEquals($expected, $query, [], 'some-id');